no more warnings

+++ src/main/java/com/lho/demo/Loop.java
@@ -20,6 +20,6 @@ public class Loop extends AnimationTimer {
     }
 
     public static double deltaTime() {
-        return (double) deltaTimeMs;
+        return deltaTimeMs;
     }
 }
